Refactor to remove fast refresh warning
