+++ src/components/screens/ListScreen/Drawer/General.tsx
@@ -1,5 +1,3 @@
-import { useKanjiInfo } from "@/kanji-worker/kanji-worker-provider";
-
 import { Table, TableBody, TableCell, TableRow } from "@/components/ui/table";
 import { Badge } from "@/components/ui/badge";
 import { Separator } from "@radix-ui/react-select";
@@ -8,6 +6,7 @@ import { BasicLoading } from "@/components/common/BasicLoading";
 import { DefaultErrorFallback } from "@/components/common/DefaultErrorFallback";
 import { ExternalKanjiLinks } from "@/components/common/ExternalKanjiLinks";
 import { GeneralKanjiItem } from "@/lib/kanji-info-types";
+import { useKanjiInfo } from "@/kanji-worker/kanji-worker-hooks";
 
 const hasData = (data?: number) => data != null && data !== -1;
