+++ src/components/screens/ListScreen/InfoCard/HiraganaWord.tsx
@@ -1,7 +1,8 @@
 import { Button } from "@/components/ui/button";
 import { useState } from "react";
-import { badgeCn, SpanBadge } from "@/components/ui/badge";
+import { SpanBadge } from "@/components/ui/badge";
 import * as wanakana from "wanakana";
+import { badgeCn } from "@/components/ui/badge-utils";
 
 const useHiraganaWordButton = (rawKana: string) => {
   const [showRomaji, setShowRomaji] = useState(false);
