+++ src/components/screens/ListScreen/InfoCard/SmallCard.tsx
@@ -1,13 +1,14 @@
 import { Badge } from "@/components/ui/badge";
-import {
-  useIsKanjiWorkerReady,
-  useKanjiInfo,
-} from "@/kanji-worker/kanji-worker-provider";
+
 import { HoverItemReturnData } from "@/lib/kanji-info-types";
 import { FrequencyBadges } from "./FrequencyBadges";
 import { JLPTBadge } from "@/components/common/JLPTBadge";
 import { BasicLoading } from "@/components/common/BasicLoading";
 import { DefaultErrorFallback } from "@/components/common/DefaultErrorFallback";
+import {
+  useIsKanjiWorkerReady,
+  useKanjiInfo,
+} from "@/kanji-worker/kanji-worker-hooks";
 
 export const SmallKanjiCard = ({ kanji }: { kanji: string }) => {
   const data = useKanjiInfo(kanji, "hover-card");
