+++ src/components/ui/badge.tsx
@@ -1,27 +1,8 @@
 import * as React from "react";
-import { cva, type VariantProps } from "class-variance-authority";
+import { type VariantProps } from "class-variance-authority";
 
 import { cn } from "@/lib/utils";
-
-const badgeVariants = cva(
-  "inline-flex items-center rounded-full border px-2.5 py-0.5 text-xs font-semibold transition-colors focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2",
-  {
-    variants: {
-      variant: {
-        default:
-          "border-transparent bg-primary text-primary-foreground hover:bg-primary/80",
-        secondary:
-          "border-transparent bg-secondary text-secondary-foreground hover:bg-secondary/80",
-        destructive:
-          "border-transparent bg-destructive text-destructive-foreground hover:bg-destructive/80",
-        outline: "text-foreground",
-      },
-    },
-    defaultVariants: {
-      variant: "default",
-    },
-  }
-);
+import { badgeVariants } from "./badge-utils";
 
 export interface BadgeProps
   extends React.HTMLAttributes<HTMLDivElement>,
@@ -38,7 +19,5 @@ function SpanBadge({ className, variant, ...props }: BadgeProps) {
     <span className={cn(badgeVariants({ variant }), className)} {...props} />
   );
 }
-const badgeCn = cn(badgeVariants({ variant: "default" }));
-const badgeCnOutline = cn(badgeVariants({ variant: "outline" }));
 
-export { Badge, badgeVariants, badgeCn, badgeCnOutline, SpanBadge };
+export { Badge, SpanBadge };
