+++ src/kanji-worker/kanji-worker-provider.tsx
@@ -1,13 +1,5 @@
-import {
-  createContext,
-  ReactNode,
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-} from "react";
+import { ReactNode, useCallback, useEffect, useRef, useState } from "react";
 import KANJI_WORKER_SINGLETON from "@/kanji-worker/kanji-worker-promise-wrapper";
-import { useContextWithCatch } from "../providers/helpers";
 import {
   GeneralKanjiItem,
   HoverItemReturnData,
@@ -19,22 +11,17 @@ import {
   VocabExtendedInfo,
 } from "@/lib/kanji-info-types";
 import { KanjiExtendedInfo, KanjiMainInfo } from "@/lib/kanji-worker-types";
-import { useSearchSettings } from "../providers/search-settings-provider";
-import { SearchSettings } from "@/lib/settings";
-
-const requestWorker = KANJI_WORKER_SINGLETON.request;
-type KanjiRequestFn = (
-  k: string,
-  type: KanjiInfoRequestType
-) => Promise<unknown>;
+import {
+  ActionContext,
+  GetBasicKanjiInfoContext,
+  IsReadyContext,
+} from "./kanji-worker-hooks";
 
 type GetBasicKanjiInfo = (kanji: string) => KanjiMainInfo | null;
 
-const ActionContext = createContext<KanjiRequestFn | null>(null);
-const IsReadyContext = createContext<boolean>(false);
-const GetBasicKanjiInfoContext = createContext<GetBasicKanjiInfo | null>(null);
+const requestWorker = KANJI_WORKER_SINGLETON.request;
 
-export const extractKanjiHoverData = (
+const extractKanjiHoverData = (
   kanjiInfo: KanjiCacheItem,
   kanjiInfoExtended: KanjiExtendedInfo & VocabExtendedInfo,
   kanjiCache?: KanjiCacheType | null,
@@ -289,164 +276,3 @@ export function KanjiWorkerProvider({
     </ActionContext.Provider>
   );
 }
-
-export const useKanjiWorkerRequest = () => {
-  const fn = useContextWithCatch(
-    ActionContext,
-    "KanjiWorker",
-    "KanjirWorkerRequest"
-  );
-  return fn;
-};
-
-export const useIsKanjiWorkerReady = () => {
-  const ready = useContextWithCatch(
-    IsReadyContext,
-    "KanjiWorker",
-    "IsKanjiWorkerReady"
-  );
-  return ready;
-};
-
-export const useGetKanjiInfoFn = () => {
-  const fn = useContextWithCatch(
-    GetBasicKanjiInfoContext,
-    "KanjiWorker",
-    "GetKanjiInfoFn"
-  );
-  return fn;
-};
-
-export type Status = "idle" | "loading" | "error" | "success";
-
-export const useKanjiSearch = (searchSettings: SearchSettings) => {
-  const [state, setState] = useState<{
-    status: Status;
-    data?: string[];
-    error?: string | null;
-  }>({ status: "idle" });
-
-  const lastRequestedSettings = useRef<null | SearchSettings>(null);
-
-  useEffect(() => {
-    const doubleRequest = lastRequestedSettings.current === searchSettings;
-
-    if (doubleRequest) {
-      return;
-    }
-
-    lastRequestedSettings.current = searchSettings;
-
-    setState((prev) => {
-      return { status: "loading", data: prev.data };
-    });
-
-    requestWorker({ type: "search", payload: searchSettings })
-      .then((result: unknown) => {
-        setState({
-          status: "success",
-          error: null,
-          data: result as string[],
-        });
-      })
-      .catch((error) => {
-        setState({ status: "error", error });
-      });
-  }, [searchSettings]);
-
-  return state;
-};
-
-export const useKanjiSearchCount = (searchSettings: SearchSettings) => {
-  const [state, setState] = useState<{
-    status: Status;
-    data?: number;
-    error?: string | null;
-  }>({ status: "idle" });
-
-  const lastRequestedSettings = useRef<null | SearchSettings>(null);
-
-  useEffect(() => {
-    const doubleRequest = lastRequestedSettings.current === searchSettings;
-
-    if (doubleRequest) {
-      return;
-    }
-
-    lastRequestedSettings.current = searchSettings;
-
-    setState((prev) => {
-      return { status: "loading", data: prev.data };
-    });
-
-    requestWorker({ type: "search-result-count", payload: searchSettings })
-      .then((result: unknown) => {
-        setState({
-          status: "success",
-          error: null,
-          data: result as number,
-        });
-      })
-      .catch((error) => {
-        setState({ status: "error", error });
-      });
-  }, [searchSettings]);
-
-  return state;
-};
-
-export const useKanjiInfo = (
-  kanji: string,
-  requestType: KanjiInfoRequestType | "none"
-) => {
-  const [state, setState] = useState<{
-    status: Status;
-    data?: unknown;
-    error?: { message: string } | null;
-  }>({
-    status: "idle",
-    data: null,
-  });
-  const requestFn = useKanjiWorkerRequest();
-
-  useEffect(() => {
-    if (requestFn == null) {
-      setState({
-        status: "error",
-        error: {
-          message: "requestFn does not exist. Please check KanjiWorkerProvider",
-        },
-      });
-
-      return;
-    }
-
-    if (requestType === "none") {
-      return;
-    }
-
-    setState((prev) => {
-      return { status: "loading", data: prev.data };
-    });
-    requestFn(kanji, requestType)
-      .then((result: unknown) => {
-        setState({
-          status: "success",
-          error: null,
-          data: result,
-        });
-      })
-      .catch((error) => {
-        setState({ status: "error", error });
-      });
-
-    return;
-  }, [kanji, requestType, requestFn]);
-  return state;
-};
-
-export const useKanjiSearchResult = () => {
-  const searchSettings = useSearchSettings();
-  const results = useKanjiSearch(searchSettings);
-  return results;
-};
