+++ src/providers/item-settings-hooks.tsx
@@ -1,0 +1,34 @@
+import { useDeferredValue } from "react";
+import { createContextComponents, useContextWithCatch } from "./helpers";
+import { ItemSettings } from "@/lib/settings";
+
+export const defaultItemSettingsValue: ItemSettings = {
+  cardType: "compact",
+  borderColorAttached: true,
+  backgroundColorSettingDataSource: "none",
+};
+
+export const itemSettings = createContextComponents<ItemSettings>(
+  defaultItemSettingsValue
+);
+
+const providerName = "ItemSettings";
+
+export function useItemSettings() {
+  const context = useContextWithCatch(itemSettings.StateContext, providerName);
+  return context;
+}
+
+export function useDeferredItemSettings() {
+  const context = useItemSettings();
+  const value = useDeferredValue(context);
+  return value;
+}
+export function useItemSettingsDispatch() {
+  const context = useContextWithCatch(
+    itemSettings.DispatchContext,
+    providerName,
+    `${providerName}Dispatch`
+  );
+  return context;
+}
